+++ common/src/main/java/dev/terminalmc/clientsort/client/ClientSort.java
@@ -16,59 +16,23 @@
 
 package dev.terminalmc.clientsort.client;
 
-import com.mojang.blaze3d.platform.InputConstants;
 import dev.terminalmc.clientsort.client.config.Config;
 import dev.terminalmc.clientsort.client.gui.ControlButtonManager;
 import dev.terminalmc.clientsort.client.network.InteractionManager;
 import dev.terminalmc.clientsort.client.order.SortOrder;
+import dev.terminalmc.clientsort.client.util.Keybinds;
 import dev.terminalmc.clientsort.util.ModLogger;
-import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.resources.ResourceLocation;
 
-import java.util.List;
-
-import static dev.terminalmc.clientsort.util.Localization.translationKey;
+import static dev.terminalmc.clientsort.client.config.Config.options;
 
 public class ClientSort {
 
     public static final String MOD_ID = dev.terminalmc.clientsort.ClientSort.MOD_ID;
     public static final String MOD_NAME = dev.terminalmc.clientsort.ClientSort.MOD_NAME;
     public static final ModLogger LOG = dev.terminalmc.clientsort.ClientSort.LOG;
-
-    public static final KeyMapping EDIT_KEY = new KeyMapping(
-            translationKey("key", "group.edit"),
-            InputConstants.Type.KEYSYM,
-            InputConstants.UNKNOWN.getValue(),
-            translationKey("key", "group")
-    );
-    public static final KeyMapping SORT_KEY = new KeyMapping(
-            translationKey("key", "group.op.sort"),
-            InputConstants.Type.MOUSE,
-            InputConstants.MOUSE_BUTTON_MIDDLE,
-            translationKey("key", "group")
-    );
-    public static final KeyMapping STACK_FILL_KEY =
-            new KeyMapping(
-                    translationKey("key", "group.op.stackFill"),
-                    InputConstants.Type.KEYSYM,
-                    InputConstants.UNKNOWN.getValue(),
-                    translationKey("key", "group")
-            );
-    public static final KeyMapping TRANSFER_KEY =
-            new KeyMapping(
-                    translationKey("key", "group.op.transfer"),
-                    InputConstants.Type.KEYSYM,
-                    InputConstants.UNKNOWN.getValue(),
-                    translationKey("key", "group")
-            );
-    public static final List<KeyMapping> KEYBINDS = List.of(
-            EDIT_KEY,
-            SORT_KEY,
-            STACK_FILL_KEY,
-            TRANSFER_KEY
-    );
 
     public static boolean searchOrderUpdated = false;
 
@@ -92,7 +56,7 @@ public class ClientSort {
         ControlButtonManager.afterScreenInit(screen);
     }
 
-    public static void onConfigSaved(Config config) {
+    public static void afterConfigSaved(Config config) {
         Config.Options options = config.options;
         // Convert config sort order strings into enum values
         options.sortOrder = SortOrder.SORT_MODES.get(options.sortOrderStr);
@@ -106,5 +70,17 @@ public class ClientSort {
         // Update class caches
         ControlButtonManager.reloadLayoutClasses(options.buttonLayouts.keySet());
         ControlButtonManager.reloadPolicyClasses(options.classPolicies.keySet());
+    }
+
+    public static void afterGameStart() {
+        if (options().deconflictKeybinds) {
+            Keybinds.configToKeybinds();
+        }
+    }
+
+    public static void afterKeyMapReset() {
+        if (!options().deconflictKeybinds) {
+            Keybinds.keybindsToConfig();
+        }
     }
 }